refactor: change naming variables

+++ ewm-main-service/src/main/java/ru/practicum/ewm/event/dto/EventShortDto.java
@@ -11,13 +11,22 @@ import ru.practicum.ewm.user.dto.UserShortDto;
 @EqualsAndHashCode
 @ToString
 public class EventShortDto {
+
     private String annotation;
+
     private CategoryDto category;
-    private Long confirmedRequests;
+
+    private Integer confirmedRequests;
+
     private String eventDate;
+
     private Long id;
+
     private UserShortDto initiator;
+
     private Boolean paid;
+
     private String title;
+
     private Long views;
 }
