+++ ewm-main-service/src/main/java/ru/practicum/ewm/request/RequestRepository.java
@@ -4,6 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
+import java.util.Map;
 
 @Repository
 public interface RequestRepository extends JpaRepository<Request, Long> {
@@ -13,9 +14,10 @@ public interface RequestRepository extends JpaRepository<Request, Long> {
 
     List<Request> findAllByEvent_IdAndStatus(long eventId, StateRequest status);
 
-    Long countAllByEvent_IdAndStatus(long eventId, StateRequest status);
+    Integer countAllByEvent_IdAndStatus(long eventId, StateRequest status);
 
     List<Request> findAllByEvent_Id(long eventId);
 
     List<Request> findAllByEvent_IdAndRequester_IdAndStatus(long eventId, long requesterId, StateRequest status);
+
 }
